new upstream pull

+++ chrome/browser/resources/password_manager/password_manager_proxy.ts
@@ -10,7 +10,7 @@
 
 import {loadTimeData} from 'chrome://resources/js/load_time_data.js';
 
-import {PageCallbackRouter, PageHandlerFactory, PageHandlerRemote} from './password_manager.mojom-webui.js';
+import {type ActorLoginPermission, PageCallbackRouter, PageHandlerFactory, PageHandlerRemote} from './password_manager.mojom-webui.js';
 
 export type BlockedSite = chrome.passwordsPrivate.ExceptionEntry;
 
@@ -437,6 +437,16 @@ export interface PasswordManagerProxy {
    * Deletes all password manager data (passwords, passkeys, etc.)
    */
   deleteAllPasswordManagerData(): Promise<boolean>;
+
+  /**
+   * Returns the list of sites that can be used for Actor Login.
+   */
+  getActorLoginPermissions(): Promise<ActorLoginPermission[]>;
+
+  /**
+   * Revokes actor login permission for all credentials matching the site.
+   */
+  revokeActorLoginPermission(site: ActorLoginPermission): void;
 }
 
 /**
@@ -534,7 +544,11 @@ export class PasswordManagerImpl implements PasswordManagerProxy {
   }
 
   showAddShortcutDialog() {
-    chrome.passwordsPrivate.showAddShortcutDialog();
+    if (!loadTimeData.getBoolean('enablePasswordManagerMojoApi')) {
+      chrome.passwordsPrivate.showAddShortcutDialog();
+      return;
+    }
+    this.handler.showAddShortcutDialog();
   }
 
   requestCredentialsDetails(ids: number[]) {
@@ -627,7 +641,11 @@ export class PasswordManagerImpl implements PasswordManagerProxy {
   }
 
   switchBiometricAuthBeforeFillingState() {
-    return chrome.passwordsPrivate.switchBiometricAuthBeforeFillingState();
+    if (!loadTimeData.getBoolean('enablePasswordManagerMojoApi')) {
+      return chrome.passwordsPrivate.switchBiometricAuthBeforeFillingState();
+    }
+    return this.handler.switchBiometricAuthBeforeFillingState().then(
+        result => result.success);
   }
 
   showExportedFileInShell(filePath: string) {
@@ -650,7 +668,11 @@ export class PasswordManagerImpl implements PasswordManagerProxy {
   }
 
   extendAuthValidity() {
-    chrome.passwordsPrivate.extendAuthValidity();
+    if (!loadTimeData.getBoolean('enablePasswordManagerMojoApi')) {
+      chrome.passwordsPrivate.extendAuthValidity();
+      return;
+    }
+    this.handler.extendAuthValidity();
   }
 
   addAccountStorageEnabledStateListener(
@@ -678,15 +700,27 @@ export class PasswordManagerImpl implements PasswordManagerProxy {
   }
 
   isAccountStorageEnabled() {
-    return chrome.passwordsPrivate.isAccountStorageEnabled();
+    if (!loadTimeData.getBoolean('enablePasswordManagerMojoApi')) {
+      return chrome.passwordsPrivate.isAccountStorageEnabled();
+    }
+    return this.handler.isAccountStorageEnabled().then(
+        result => result.enabled);
   }
 
   setAccountStorageEnabled(enabled: boolean) {
-    chrome.passwordsPrivate.setAccountStorageEnabled(enabled);
+    if (!loadTimeData.getBoolean('enablePasswordManagerMojoApi')) {
+      chrome.passwordsPrivate.setAccountStorageEnabled(enabled);
+      return;
+    }
+    this.handler.setAccountStorageEnabled(enabled);
   }
 
   shouldShowAccountStorageSettingToggle() {
-    return chrome.passwordsPrivate.shouldShowAccountStorageSettingToggle();
+    if (!loadTimeData.getBoolean('enablePasswordManagerMojoApi')) {
+      return chrome.passwordsPrivate.shouldShowAccountStorageSettingToggle();
+    }
+    return this.handler.shouldShowAccountStorageSettingToggle().then(
+        result => result.shouldShow);
   }
 
   movePasswordsToAccount(ids: number[]) {
@@ -698,11 +732,19 @@ export class PasswordManagerImpl implements PasswordManagerProxy {
   }
 
   changePasswordManagerPin() {
-    return chrome.passwordsPrivate.changePasswordManagerPin();
+    if (!loadTimeData.getBoolean('enablePasswordManagerMojoApi')) {
+      return chrome.passwordsPrivate.changePasswordManagerPin();
+    }
+    return this.handler.changePasswordManagerPin().then(
+        result => result.success);
   }
 
   isPasswordManagerPinAvailable() {
-    return chrome.passwordsPrivate.isPasswordManagerPinAvailable();
+    if (!loadTimeData.getBoolean('enablePasswordManagerMojoApi')) {
+      return chrome.passwordsPrivate.isPasswordManagerPinAvailable();
+    }
+    return this.handler.isPasswordManagerPinAvailable().then(
+        result => result.isAvailable);
   }
 
   disconnectCloudAuthenticator() {
@@ -714,10 +756,19 @@ export class PasswordManagerImpl implements PasswordManagerProxy {
   }
 
   deleteAllPasswordManagerData() {
-    return loadTimeData.getBoolean('enablePasswordManagerMojoApi') ?
-        this.handler.deleteAllPasswordManagerData().then(
-            result => result.success) :
-        chrome.passwordsPrivate.deleteAllPasswordManagerData();
+    if (!loadTimeData.getBoolean('enablePasswordManagerMojoApi')) {
+      return chrome.passwordsPrivate.deleteAllPasswordManagerData();
+    }
+    return this.handler.deleteAllPasswordManagerData().then(
+        result => result.success);
+  }
+
+  getActorLoginPermissions() {
+    return this.handler.getActorLoginPermissions().then(result => result.sites);
+  }
+
+  revokeActorLoginPermission(site: ActorLoginPermission) {
+    this.handler.revokeActorLoginPermission(site);
   }
 
   static getInstance(): PasswordManagerProxy {
